% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_coeffs.R
\name{avgEst}
\alias{avgEst}
\title{Weighted Average of Model Estimates}
\usage{
avgEst(e, weights = "equal", est.names = NULL, ...)
}
\arguments{
\item{e}{A list or nested list of numeric vectors, comprising the model
estimates. In the latter case, these should correspond to estimates for
candidate models for each of a set of different response variables.}

\item{weights}{An optional numeric vector of weights to use for model
averaging, or a named list of such vectors. The former should be supplied
when \code{m} is a list, and the latter when it is a nested list (with
matching list names). If \code{weights = "equal"} (default), a simple
average is calculated instead.}

\item{est.names}{An optional character vector of names used to extract and/or
sort estimates from the output.}

\item{...}{Not used.}
}
\value{
A numeric vector of the model-averaged estimates, or a list of such
  vectors.
}
\description{
Calculate a weighted average of model estimates (e.g.
  coefficients, fitted values, residuals) for a set of models.
}
\details{
This function can be used to calculate a weighted average of model
  estimates such as coefficients, fitted values, or residuals, where models
  are typically competing candidate models fit to the same response variable
  and using the same data. Weights are typically a 'weight of evidence' type
  metric such as Akaike model weights (Burnham & Anderson 2002, Burnham
  \emph{et al.} 2011), which can be calculated in \code{R} using packages
  such as \pkg{MuMIn} or \pkg{AICcmodavg}. However, weights of any sort can
  be used, provided they are numeric. If none are supplied, a standard
  average is calculated instead.

  Averaging is performed via the 'full/zero' rather than
  'subset/conditional/natural' method, meaning that zero is substituted for
  any missing estimats (coefficients) prior to calculations - providing a
  form of 'shrinkage' for estimates (Burnham & Anderson 2002, Grueber
  \emph{et al.} 2011).
}
\references{
Burnham, K. P., & Anderson, D. R. (2002). \emph{Model Selection
  and Multimodel Inference: A Practical Information-Theoretic Approach} (2nd
  ed.). New York: Springer-Verlag. Retrieved from
  \url{https://www.springer.com/gb/book/9780387953649}

  Burnham, K. P., Anderson, D. R., & Huyvaert, K. P. (2011). AIC model
  selection and multimodel inference in behavioral ecology: some background,
  observations, and comparisons. \emph{Behavioral Ecology and Sociobiology},
  \strong{65}(1), 23-35. \url{https://doi.org/10.1007/S00265-010-1029-6}

  Dormann, C. F., Calabrese, J. M., Guillera-Arroita, G., Matechou, E., Bahn,
  V., Barton, K., ... Hartig, F. (2018). Model averaging in ecology: a review
  of Bayesian, information-theoretic, and tactical approaches for predictive
  inference. \emph{Ecological Monographs}, \strong{88}(4), 485-504.
  \url{https://doi.org/10.1002/ECM.1309}

  Grueber, C. E., Nakagawa, S., Laws, R. J., & Jamieson, I. G. (2011).
  Multimodel inference in ecology and evolution: challenges and solutions:
  Multimodel inference. \emph{Journal of Evolutionary Biology}, \code{24}(4),
  699-711. \url{https://doi.org/10.1111/j.1420-9101.2010.02210.x}
}
\seealso{
\code{\link[stats]{weighted.mean}}
}
