% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_functions.R
\name{bootCI}
\alias{bootCI}
\title{Bootstrap Confidence Intervals}
\usage{
bootCI(m, conf = 0.95, type = "bca", digits = 3, bci.arg = NULL,
  ...)
}
\arguments{
\item{m}{A fitted model object of class \code{lm}, \code{glm}, or
\code{merMod}, or a named list/nested list of such objects. \code{m} can
also be a boot object(s)/list (class \code{"boot"}), corresponding to
bootstrapped estimates from fitted models.}

\item{conf}{A numeric value specifying the confidence level for the
intervals.}

\item{type}{The type of confidence interval to return (defaults to
\code{"bca"} - see Details). See \code{\link[boot]{boot.ci}} for further
specification details.}

\item{digits}{The number of significant digits to return for numeric values.}

\item{bci.arg}{A named list of additional arguments to \code{boot.ci}, which
should not include argument \code{index}.}

\item{...}{Arguments to \code{bootSEM}.}
}
\value{
A data frame of the original estimates and the bootstrapped
  confidence intervals, or a list of same.
}
\description{
Calculate confidence intervals using bootstrapped model
  estimates.
}
\details{
This function is essentially a wrapper for \code{boot.ci} from the
  \pkg{boot} package, which will return confidence intervals of the specified
  type and level calculated from bootstrapped model estimates. If a model or
  models are supplied to \code{m}, bootstrapping will first be performed via
  \code{bootSEM}. Estimates for which the confidence intervals do not contain
  zero are highlighted with a star.

  Nonparametric bias-corrected and accelerated confidence intervals
  (BC\emph{a}, Efron 1987) are calculated by default, which provide the most
  accurate coverage across the widest variety of bootstrap sampling
  distributions (Puth \emph{et al.} 2015). They will, however, be
  inappropriate for parametric resampling - due to the requirement for
  empirical influence values calculated from the data
  (\url{http://bit.ly/2UzmA7Y}) - in which case the default will be set to
  the boostrap percentile method instead (\code{"perc"}).

  NOTE: All bootstrapped confidence intervals will underestimate the true
  nominal coverage to some extent when sample size is small (Chernick &
  Labudde 2009), so the appropriate caution should be exercised in
  interpretation in these cases. Comparison of different interval types may
  be informative. For example, normal-theory based intervals may outperform
  bootstrap percentile methods when n < 34 (Hesterberg 2015). Ultimately
  though, the bootstrap is not a solution to small sample size
  (\url{http://bit.ly/2GpfLMn}).
}
\examples{
## CI's from bootstrapped SEM
bootCI(Shipley.SEM.boot)

## From original SEM
## (not usually recommended - better to use saved boot objects)

\dontrun{

system.time(
  Shipley.SEM.CI <- bootCI(Shipley.SEM, ran.eff = "site", seed = 53908)
)
}
}
\references{
Chernick, M. R., & Labudde, R. A. (2009). Revisiting Qualms about
  Bootstrap Confidence Intervals. \emph{American Journal of Mathematical and
  Management Sciences}, \strong{29}(3–4), 437–456. \url{https://doi.org/c8zv}

  Efron, B. (1987). Better Bootstrap Confidence Intervals. \emph{Journal of
  the American Statistical Association}, \strong{82}(397), 171–185.
  \url{https://doi.org/gfww2z}

  Hesterberg, T. C. (2015). What Teachers Should Know About the Bootstrap:
  Resampling in the Undergraduate Statistics Curriculum. \emph{The American
  Statistician}, \strong{69}(4), 371–386. \url{https://doi.org/gd85v5}

  Puth, M.-T., Neuhäuser, M., & Ruxton, G. D. (2015). On the variety of
  methods for calculating confidence intervals by bootstrapping.
  \emph{Journal of Animal Ecology}, \strong{84}(4), 892–897.
  \url{https://doi.org/f8n9rq}
}
\seealso{
\code{\link[boot]{boot.ci}}, \code{\link[semEff]{bootSEM}}
}
