% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{Class}
\alias{Class}
\alias{isList}
\alias{isBoot}
\alias{isMod}
\alias{isGlm}
\alias{isMerMod}
\title{Determine Class of Objects}
\usage{
isList(x)

isBoot(x)

isMod(x)

isGlm(x)

isMerMod(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
A logical value.
}
\description{
Functions to determine the class of an R object.
}
\section{Functions}{
\itemize{
\item \code{isList}: Is object of class "list"?

\item \code{isBoot}: Is object a boot object (output of function \code{boot})?

\item \code{isMod}: Is object a linear or generalised linear (mixed) model?

\item \code{isGlm}: Is object a generalised linear (mixed) model?

\item \code{isMerMod}: Is object a mixed model (of class \code{merMod})?
}}

