% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_eff_functions.R
\name{Effects}
\alias{Effects}
\alias{effects}
\alias{direct}
\alias{indirect}
\alias{total}
\title{Extract SEM Effects}
\usage{
effects(x, type = "orig", responses = NULL)

direct(...)

indirect(...)

total(...)
}
\arguments{
\item{x}{An object of class "semEff".}

\item{type}{The type of effects to return. Must be one of \code{c("orig",
"boot")} (defaults to \code{"orig"}).}

\item{responses}{A vector of names of response variables for which to return
effects. If \code{NULL} (default), all responses are used.}
}
\value{
A list containing the original or bootstrapped effects for each
  response variable, as numeric vectors or matrices (respectively).
}
\description{
Extract direct, indirect, or total effects for a piecewise SEM.
}
\details{
These are simple extractor functions for effects calculated using
  \code{semEff}, intended for convenience.
}
\section{Functions}{
\itemize{
\item \code{effects}: Extract all effects.

\item \code{direct}: Extract direct effects.

\item \code{indirect}: Extract indirect effects.

\item \code{total}: Extract total effects.
}}

