% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_coeffs.R
\name{getData}
\alias{getData}
\title{Model Data}
\usage{
getData(m, subset = FALSE, merge = FALSE, ...)
}
\arguments{
\item{m}{A fitted model object of class \code{lm}, \code{glm}, or
\code{merMod}, or a list or nested list of such objects.}

\item{subset}{Logical. If \code{TRUE}, rows with missing observations
(\code{NA}) are removed (if those variable(s) were used to fit model(s)).}

\item{merge}{Logical. If \code{TRUE}, and \code{m} is a list or nested list,
all unique data columns are merged into one data frame, in the order in
which they are encountered.}

\item{...}{Arguments to \code{eval}.}
}
\value{
A data frame of the variables used to fit the model(s).
}
\description{
Extract the data used to fit a model.
}
\details{
This is a simple function to extract the data used to fit a model
  by evaluating the data slot of the model call object. If the model was fit
  using separate variables (i.e. vectors) instead of a data frame, data is
  returned via \code{model.frame} instead (with a warning). If \code{m} is a
  list and \code{merge = TRUE}, a single dataset containing all variables
  used to fit models is returned.
}
\examples{

}
\seealso{
\code{\link[base]{eval}}, \code{\link[stats]{getCall}},
  \code{\link[stats]{model.frame}}
}
