% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdEff-fun.R
\name{getFamily}
\alias{getFamily}
\title{Get Model Error Distribution Family}
\usage{
getFamily(mod)
}
\arguments{
\item{mod}{A fitted model object, or a list or nested list of such objects.}
}
\value{
A model \code{"family"} object, or a list or nested list of such objects.
}
\description{
Extract the error distribution family (and link function) for a
fitted model as a \code{"family"} object.
}
\details{
\code{getFamily()} returns a family object for a range of different model
classes, similar to \code{\link[=family]{family()}}. However, it can be also be used for some
without an existing family method. Mostly for internal use.
}
\examples{
# SEM model error distributions
getFamily(shipley.sem)
}
