% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_coeffs.R
\name{getY}
\alias{getY}
\title{Model Response Variable}
\usage{
getY(m, family = NULL, data = NULL, link = FALSE, ...)
}
\arguments{
\item{m}{A fitted model object of class \code{lm}, \code{glm}, or
\code{merMod}, or alternatively a numeric vector corresponding to a
variable to be transformed. \code{m} can also be a list or nested list of
such objects.}

\item{family}{A character string specifying the GLM error family (e.g.
"binomial", "poisson"). Ignored if \code{m} is a fitted model(s).}

\item{data}{An optional dataset used to first re-fit the model(s) (if
\code{m} is a model object or list).}

\item{link}{Logical. If \code{TRUE}, return the (GLM) response variable on
the link scale (see Details).}

\item{...}{Not used.}
}
\value{
A numeric vector corresponding to the response variable in the link
  scale, or an array, list or nested list of such vectors.
}
\description{
Extract the response variable from a fitted model in original or
  link scale (for GLM's).
}
\details{
This function will reconstruct the response variable from a model by
  summing the fitted values and the response residuals. If \code{link =
  TRUE}, the response is then transformed using the GLM link function, and,
  if any values are undefined, the 'working' response variable of the model
  is returned instead (see below). The function can also be used to transform
  a variable (supplied to \code{m}) using the default link function of the
  specified \code{family} - in which case the \code{link} argument is
  ignored.

  \strong{Use of the Working Response Variable}

  A key challenge in generating fully standardised model coefficients for a
  generalised linear model (GLM) with a non-gaussian link function is the
  (in)ability to calculate appropriate standardised ranges (typically the
  standard deviation) for the response variable in the link scale. This is
  because the directly-transformed response will often have undefined values.
  Although methods for circumventing this issue by indirectly estimating the
  standard deviation of the response in the link scale have previously been
  proposed - including a general variance-based method using a pseudo
  R-squared (Menard 2011) and another using a latent-theoretic framework for
  binomial models (Grace \emph{et al.} 2018) - here another approach is used.
  Where any transformed values are undefined, the function will instead
  return the synthetic 'working' response from the final iteration of the
  iteratively reweighted least squares (IRLS) algorithm of the GLM (McCullagh
  & Nelder 1989). This is reconstructed as the sum of the linear predictor
  and the working residuals. The advantage of this approach is that a
  relatively straightforward 'transformation' of any non-gaussian response is
  readily attainable in all cases. The standard deviation (or other relevant
  range) can then be calculated using values of the transformed response and
  used to scale the coefficients. An additional benefit for piecewise SEM's
  is that the transformed rather than original response can be used as a
  predictor in other models, ensuring that standardised indirect and total
  effects are calculated correctly (i.e. using a common standard deviation).

  As the working response from a well-fitting model will be more 'accurate' -
  in the sense that the inverse-transformed values will more closely resemble
  the original response - the function always calculates it from a new GLM
  fit with the original response as the sole predictor (and using the same
  error family). Using the response as the predictor is a convenient way to
  ensure a near-optimal model fit, one which is limited only by the extent of
  non-linearity in the relationship between the response and its
  link-transformed expected values. The inverse-transformation of the working
  response from this model consistently produces an almost perfect
  correlation with the original response (see examples). This approach also
  ensures that the transformed values, and hence the standard deviation, are
  the same for any GLM fitted to the same response variable (provided it uses
  the same link function), and so facilitates parameter comparisons across
  different models.

  The solution here of using the working response variable from the GLM to
  scale coefficients is primarily a practical one, but would generally be
  classed as an 'observed-empirical' approach to standardisation by Grace
  \emph{et al.} (2018) - as it utilises model error variance (i.e. the
  working residuals) rather than theoretical distribution-specific variance.
}
\references{
Grace, J.B., Johnson, D.J., Lefcheck, J.S. and Byrnes, J.E.K.
  (2018) Quantifying relative importance: computing standardized effects in
  models with binary outcomes. \emph{Ecosphere} \strong{9}, e02283.
  \url{https://doi.org/10.1002/ECS2.2283}

  McCullagh P. and Nelder, J. A. (1989) \emph{Generalized Linear Models} (2nd
  Edition). London: Chapman and Hall.

  Menard, S. (2011) Standards for Standardized Logistic Regression
  Coefficients. \emph{Social Forces} \strong{89}, 1409-1428.
  \url{https://doi.org/10.1093/SF/89.4.1409}
}
