% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_coeff_fun.R
\name{glt}
\alias{glt}
\title{Generalised Link Transformation}
\usage{
glt(x, family = NULL)
}
\arguments{
\item{x}{a positive numeric vector, corresponding to a variable to be
transformed. Can also be a list or nested list of such objects.}

\item{family}{Optional, the error distribution family containing the link
function which will be used to transform \code{x} (see
\code{\link[stats]{family}} for specification details).}
}
\value{
A numeric vector of the transformed values, or an array, list of
  vectors/arrays, or nested list.
}
\description{
Transform a numeric vector using a GLM link function.
}
\details{
\code{glt(x)} is a shortcut to \code{getY(x, family = family)},
  where \code{x} is a positive numeric vector and \code{family} is its error
  distribution family. If this is not specified (default), then it is
  determined (roughly) from \code{x}, with \code{binomial(link = "logit")}
  used when all x <= 1 and \code{poisson(link = "log")} otherwise. If the
  transformation results in undefined values, an estimate based on the
  'working' response variable of the GLM is returned instead (see
  \code{\link[semEff]{getY}}).

  Although \code{glt} is generally intended for binomial or poisson
  variables, any variable which can be fit using \code{glm} can be supplied.
}
\seealso{
\code{\link[stats]{glm}}
}
