% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootEff-fun.R
\name{print.bootCI}
\alias{print.bootCI}
\title{Print \code{"bootCI"} Objects}
\usage{
\method{print}{bootCI}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"bootCI"}.}

\item{...}{Further arguments passed to or from other methods. Not currently
used.}
}
\value{
A summary table of the effects and bootstrapped confidence intervals
(data frame).
}
\description{
A \code{\link[=print]{print()}} method for an object of class \code{"bootCI"}.
}
