% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semEff-fun.R
\name{print.unicode_df}
\alias{print.unicode_df}
\title{Print Data Frames (Unicode Support)}
\usage{
\method{print}{unicode_df}(
  x,
  ...,
  digits = NULL,
  quote = FALSE,
  right = TRUE,
  row.names = TRUE,
  max = NULL
)
}
\arguments{
\item{x}{An object of class \code{"unicode_df"}.}

\item{..., digits, quote, right, row.names, max}{See \code{\link[=print.data.frame]{print.data.frame()}}.}
}
\description{
A \code{\link[=print]{print()}} method for data frames of class \code{"unicode_df"}.
}
\details{
This is a simple modification of \code{print.data.frame()} to facilitate
correct printing of unicode characters in all cases, by bypassing
\code{\link[=format.data.frame]{format.data.frame()}} (\href{https://stat.ethz.ch/pipermail/r-devel/2015-May/071252.html}{bug details},
workaround adapted from
\href{https://stat.ethz.ch/pipermail/r-devel/2015-May/071259.html}{here}).
}
