% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdEff-fun.R
\name{RVIF}
\alias{RVIF}
\title{Root Variance Inflation Factors}
\usage{
RVIF(...)
}
\arguments{
\item{...}{Arguments to \code{\link[semEff]{VIF}}.}
}
\value{
A numeric vector of the RVIFs, or an array, list of vectors/arrays,
  or nested list.
}
\description{
Calculate root variance inflation factors (RVIF) for terms in a
  fitted model(s), i.e. the square root of the (generalised) VIFs.
}
\details{
RVIFs quantify the inflation of estimate standard errors due to
  multicollinearity among predictors, and also of estimates themselves
  compared to the 'unique' (residualised) effects. RVIFs may often be more
  practical than VIFs for assessing multicollinearity, relating more directly
  to the parameters of interest.
}
