% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_coeff_functions.R
\name{sdW}
\alias{sdW}
\title{Weighted Standard Deviation}
\usage{
sdW(...)
}
\arguments{
\item{...}{Arguments to \code{varW}.}
}
\value{
A numeric value, the weighted standard deviation of \code{x}.
}
\description{
Calculate the weighted standard deviation of \code{x}.
}
\details{
This is a simple wrapper for \code{varW}, which applies the square
  root to the output.
}
\seealso{
\code{\link[stats]{sd}}, \code{\link[semEff]{varW}}
}
