% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semEff-package.R
\docType{package}
\name{semEff-package}
\alias{semEff-package}
\alias{_PACKAGE}
\title{semEff: Automatic Calculation of Effects for Piecewise Structural Equation Models}
\description{
Provides functionality to automatically calculate direct, indirect, 
    and total effects from piecewise structural equation models, comprising 
    lists of fitted models representing structured equations (with local 
    estimation). Confidence intervals are provided via bootstrapping. See 
    Shipley (2000) <doi:10.1207/s15328007sem0702_4>, Shipley (2009) 
    <doi:10.1890/08-1034.1>, and Lefcheck (2016) <doi:10.1111/2041-210x.12512>) 
    for background.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/murphymv/semEff}
  \item Report bugs at \url{https://github.com/murphymv/semEff/issues}
}

}
\author{
\strong{Maintainer}: Mark Murphy \email{murphymv@gmail.com}

}
\keyword{internal}
