% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Shipley.SEM.boot}
\alias{Shipley.SEM.boot}
\title{Bootstrapped SEM Estimates}
\format{A list of objects of class \code{"boot"}, representing bootstrapped
  estimates from fitted models.}
\usage{
Shipley.SEM.boot
}
\description{
Boostrapped estimates generated from the hypothesised SEM from
  Shipley (2009) using \code{bootSEM}, comprising 1,000 replications.
}
\examples{
## Specification

\dontrun{

Shipley.SEM.boot <- bootSEM(Shipley.SEM, ran.eff = "site", R = 1000,
                            seed = 53908)
}
}
\references{
Shipley, B. (2009). Confirmatory path analysis in a generalized
  multilevel context. \emph{Ecology}, \strong{90}(2), 363-368.
  \url{https://doi.org/10.1890/08-1034.1}
}
\keyword{datasets}
