% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Shipley.SEM.eff}
\alias{Shipley.SEM.eff}
\title{SEM Effects}
\format{A list object of class \code{"semEff"}, containing all SEM effects
  and summary tables.}
\usage{
Shipley.SEM.eff
}
\description{
SEM effects calculated from bootstrapped estimates of the
  hypothesised SEM from Shipley (2009), using \code{semEff}.
}
\examples{
## Specification

\dontrun{

Shipley.SEM.eff <- semEff(Shipley.SEM.boot)
}
}
\references{
Shipley, B. (2009). Confirmatory path analysis in a generalized
  multilevel context. \emph{Ecology}, \strong{90}(2), 363-368.
  \url{https://doi.org/10.1890/08-1034.1}
}
\keyword{datasets}
