% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shipley.sem}
\alias{shipley.sem}
\title{Hypothesised SEM from Shipley (2009)}
\format{
A list of fitted mixed models of class \code{"lmer"} and \code{"glmer"},
representing structured equations.
}
\usage{
shipley.sem
}
\description{
Hypothesised SEM from Shipley (2009)
}
\examples{
# Specification
# shipley.sem <- list(
#   DD = lme4::lmer(DD ~ lat + (1 | site) + (1 | tree), data = shipley),
#   Date = lme4::lmer(Date ~ DD + (1 | site) + (1 | tree), data = shipley),
#   Growth = lme4::lmer(Growth ~ Date + (1 | site) + (1 | tree),
#                       data = shipley),
#   Live = lme4::glmer(Live ~ Growth + (1 | site) + (1 | tree), binomial,
#                      data = shipley)
# )
}
\references{
Shipley, B. (2009). Confirmatory path analysis in a generalized
multilevel context. \emph{Ecology}, \strong{90}(2), 363-368. \url{https://doi.org/bqd43d}
}
\keyword{datasets}
