% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semEff-fun.R
\name{summary.semEff}
\alias{summary.semEff}
\title{Summarise SEM Effects}
\usage{
\method{summary}{semEff}(x, responses = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{"semEff"}.}

\item{responses}{An optional character vector, the names of one or more SEM
response variables for which to return summary tables. \code{"Correlated Errors"} can also be specified/included here (where applicable).}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
A summary method for an object of class \code{"semEff"}, which acts
the same as \code{\link[=print.semEff]{print.semEff()}} except that individual summary tables for
different response variables can be extracted.
}
