% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_coeffs.R
\name{varW}
\alias{varW}
\title{Weighted Variance}
\usage{
varW(x, w = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{w}{A numeric vector of weights of the same length as \code{x}.}

\item{na.rm}{Logical, should NA's in \code{x} be removed?}

\item{...}{Not used.}
}
\value{
A numeric value, the weighted variance of \code{x}.
}
\description{
Calculate the weighted variance of \code{x}.
}
\details{
This function calculates the weighted variance of \code{x} via the
  weighted covariance matrix (\code{cov.wt}). If no weights are supplied, the
  simple variance is returned instead. As in \code{weighted.mean},
  \code{NA}'s in \code{w} are not handled specially and will return \code{NA}
  as result.
}
\examples{
varW(1:10, w = 1:10)
varW(1:10)  # simple var

## NA handling
varW(c(1:9, NA), w = 1:10, na.rm = TRUE)  # NA in x (removed)
varW(c(1:9, NA), w = 1:10, na.rm = FALSE)  # NA in x (NA returned)
varW(1:10, w = c(1:9, NA))  # NA in w (NA returned)
}
\seealso{
\code{\link[stats]{var}}, \code{\link[stats]{cov.wt}},
  \code{\link[stats]{weighted.mean}}
}
