% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_eff_functions.R
\name{Effects}
\alias{Effects}
\alias{direct}
\alias{indirect}
\alias{total}
\title{Extract SEM Effects}
\usage{
Effects(sem.eff, type = "orig", responses = NULL)

direct(...)

indirect(...)

total(...)
}
\arguments{
\item{sem.eff}{An object of class \code{"semEff"}.}

\item{type}{The type of effects to return. Must be either \code{"orig"}
(default) or \code{"boot"}.}

\item{responses}{A character vector of names of response variables for which
to return effects. If \code{NULL} (default), all responses are used.}

\item{...}{Arguments (above) to be passed to \code{Effects}.}
}
\value{
A list containing the original or bootstrapped effects for each
  response variable, as numeric vectors or matrices (respectively).
}
\description{
Extract SEM direct, indirect, or total effects from an object of
  class \code{"semEff"}.
}
\details{
These are simple extractor functions for effects calculated using
  \code{semEff}, intended for convenience (e.g. for use in \code{predEff}).
}
\section{Functions}{
\itemize{
\item \code{Effects}: Extract all effects.

\item \code{direct}: Extract direct effects.

\item \code{indirect}: Extract indirect effects.

\item \code{total}: Extract total effects.
}}

