% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdEff-fun.R
\name{getFamily}
\alias{getFamily}
\title{Get Model Error Distribution Family}
\usage{
getFamily(mod)
}
\arguments{
\item{mod}{A fitted model object, or a list or nested list of such objects.}
}
\value{
A model \code{\link[=family]{family()}} object, or a list or nested list of such objects.
}
\description{
Extract the error distribution family for a fitted model as a
\code{\link[=family]{family()}} object.
}
\details{
\code{getFamily()} will return a(n appropriate) family object for a range
of different model classes, even those without an existing family method.
}
\examples{
# SEM model error distributions
getFamily(shipley.sem)
}
