% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{Object.Type}
\alias{Object.Type}
\alias{isList}
\alias{isBoot}
\alias{isMod}
\alias{isGls}
\alias{isBet}
\alias{isGlm}
\alias{isMerMod}
\title{Object Types}
\usage{
isList(x)

isBoot(x)

isMod(x)

isGls(x)

isBet(x)

isGlm(x)

isMerMod(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
A logical value.
}
\description{
Functions to determine the 'type' of an R object using classes.
  Intended largely for convenience and for internal use.
}
\section{Functions}{
\itemize{
\item \code{isList}: Is object a list (class \code{"list"})?

\item \code{isBoot}: Is object a boot object?

\item \code{isMod}: Is object a fitted model?

\item \code{isGls}: Is object a generalised least squares model?

\item \code{isBet}: Is object a beta regression model?

\item \code{isGlm}: Is object a generalised linear model (i.e. uses a
link function)?

\item \code{isMerMod}: Is object a mixed model (class \code{"merMod"})?
}}

\keyword{internal}
