% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{Param.Type}
\alias{Param.Type}
\alias{isInt}
\alias{isInx}
\alias{isR2}
\title{Determine Model Parameter Types}
\usage{
isInt(x)

isInx(x)

isR2(x)
}
\arguments{
\item{x}{A character vector of parameter names (e.g. names of coefficients
from \code{\link[stats]{coef}} or \code{\link[semEff]{stdCoeff}}).}
}
\value{
A logical vector of the same length as \code{x}.
}
\description{
Functions to determine the presence/absence of certain model
  parameters using names.
}
\section{Functions}{
\itemize{
\item \code{isInt}: Is parameter an intercept?

\item \code{isInx}: Is parameter a variable interaction (product term)?

\item \code{isR2}: Is parameter an R-squared value?
}}

