% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_eff_functions.R
\name{predEff}
\alias{predEff}
\title{SEM Effect Predictions}
\usage{
predEff(m, newdata = NULL, effects = NULL, eff.boot = NULL,
  type = "link", interaction = NULL, seed = 1, ci.conf = 0.95,
  ci.type = "bca", digits = 3, bci.arg = NULL, parallel = "no",
  ncpus = NULL, cl = NULL, ...)
}
\arguments{
\item{m}{A fitted model object of class \code{lm}, \code{glm}, or
\code{merMod}, or a list or nested list of such objects.}

\item{newdata}{An optional data frame of new values to predict from, which
should contain all the variables used to fit \code{m}, or all those named
in \code{effects}.}

\item{effects}{A numeric vector of effects to predict, or a list or nested
list of such vectors.}

\item{eff.boot}{A matrix of bootstrapped effects used to calculate confidence
intervals for predictions, or a list or nested list of such matrices.}

\item{type}{The type of prediction to return (for GLM's). Can be either
\code{"link"} (default) or \code{"response"}.}

\item{interaction}{An optional name of an interactive effect, for which to
return standardised effects for the main (continuous) variable across
different values or levels of interacting variables (see Details). The name
should be of the form \code{"a:b:..."}, containing all the variables
involved and matching the name of an interactive effect present in model(s)
coefficients or in \code{effects}.}

\item{seed}{Integer, seed for the random number generator (for reproducible
results). This should be set to the same value originally set for
boostrapping effects.}

\item{ci.conf}{A numeric value specifying the confidence level for confidence
intervals on predictions (and optional interactive effects).}

\item{ci.type}{The type of confidence interval to return. Can be one of
\code{c("norm","basic", "stud", "perc", "bca")} (see
\code{\link[boot]{boot.ci}}). Defaults to \code{"bca"} (see Details).}

\item{digits}{The number of significant digits to return for numeric values
(for interactive effects only).}

\item{bci.arg}{A named list of additional arguments to \code{boot.ci}, which
should not include argument \code{index}.}

\item{parallel}{The type of parallel processing to use for calculating
confidence intervals on predictions. Can be one of \code{"snow"},
\code{"multicore"}, or \code{"no"} (for none - the default).}

\item{ncpus}{Integer, number of system cores to use for parallel processing.
If \code{NULL} (default), all available cores are used.}

\item{cl}{Optional cluster to use if \code{parallel = "snow"}. If \code{NULL}
(default), a local cluster is created using the specified number of cores.}

\item{...}{Arguments to \code{stdCoeff}.}
}
\value{
A numeric vector of the predictions; or, if bootstrapped effects are
  supplied, a list containing the predictions and the upper and lower
  confidence intervals. Optional interactive effects may also be appended to
  the output.
}
\description{
Generate predicted values for SEM direct, indirect, or total
  effects on a response variable.
}
\details{
This function can be used to generate predicted values for SEM
  direct, indirect, or total effects on a response variable (arising via
  function \code{semEff}), which should be supplied to argument
  \code{effects}. The standard prediction formula is used, except that
  effects are used in place of model coefficients, with values for predictors
  drawn from the original model data (or from \code{newdata}). Fully
  standardised effects will be assumed, but if this is not the case then the
  same standardisation options as were applied originally (via
  \code{stdCoeff}) should be specified to the function. If no effects are
  supplied, standardised coefficients will be calculated and used to generate
  predictions (these will equal the model(s) fitted values if \code{unique.x
  = F} and \code{newdata = NULL}). In addition, model averaged predictions
  can be generated if \code{m} is a list and \code{weights} are supplied
  (passed to \code{stdCoeff}) - but only if no effects are supplied.

  If bootstrapped effects are also supplied (to \code{eff.boot}),
  bootstrapped predictions are calculated by predicting from each effect.
  Confidence intervals can then be returned, for which the \code{type} should
  be appropriate for the original form of bootstrap sampling (defaults to
  \code{"bca"}). If the number of observations to predict is very large,
  parallel processing may speed up the calculation of intervals.

  Predictions are always returned in the original (typically unstandardised)
  units of the (link-)response variable. For GLM's, they can be converted to
  the response scale if \code{type = "response"}.

  If the name of an interactive effect is supplied to \code{interaction},
  standardised effects (and confidence intervals) can be returned for
  predictions of a continuous 'main' variable across specified values or
  levels of interacting variable(s). The values of the variables to predict
  must be supplied to \code{newdata}, with the continuous variable being
  identified automatically as having the most unique values.
}
\seealso{
\code{\link[semEff]{semEff}}, \code{\link[semEff]{stdCoeff}},
  \code{\link[semEff]{bootCI}}, \code{\link[semEff]{pSapply}}
  \code{\link[stats]{predict}}
}
