% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_eff_functions.R
\name{print.semEff}
\alias{print.semEff}
\title{Print SEM Effects}
\usage{
\method{print}{semEff}(x)
}
\arguments{
\item{x}{An object of class "semEff".}
}
\description{
A print method for an object of class "semEff", which returns
  summary tables of effects and confidence intervals for all responses.
}
