% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semEff-package.R
\docType{package}
\name{semEff-package}
\alias{semEff}
\alias{semEff-package}
\title{semEff: Automatic Calculation of Effects for Piecewise SEM's}
\description{
This package provides functionality to automatically calculate 
    direct, indirect, and total effects from piecewise Structual Equation Models 
    (SEM), comprising lists of fitted models representing structured equations 
    (currently supported classes are "lm", "glm", "lmerMod", and "glmerMod"). 
    Confidence intervals for all effects are provided via bootstrapping.
}
\author{
\strong{Maintainer}: Mark Murphy \email{murphymv@gmail.com}

}
\keyword{internal}
