% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semEff-package.R
\docType{package}
\name{semEff-package}
\alias{semEff-package}
\alias{_PACKAGE}
\title{semEff: Automatic Calculation of Effects for Piecewise Structural Equation Models}
\description{
Provides functionality to automatically calculate direct, indirect, 
    and total effects from 'piecewise' Structural Equation Models, comprising 
    lists of fitted models representing structured equations. Confidence 
    intervals are provided via bootstrapping. Currently supported model classes 
    are "lm", "glm", and "merMod".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/murphymv/semEff}
  \item Report bugs at \url{https://github.com/murphymv/semEff/issues}
}

}
\author{
\strong{Maintainer}: Mark Murphy \email{murphymv@gmail.com}

}
\keyword{internal}
