% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semEff-package.R
\docType{package}
\name{semEff-package}
\alias{semEff-package}
\alias{_PACKAGE}
\title{semEff: Automatic Calculation of Effects for Piecewise Structural Equation Models}
\description{
Automatically calculate direct, indirect, and total effects for piecewise structural equation models, comprising lists of fitted models representing structured equations (Lefcheck, 2016 <doi:10/f8s8rb>). Confidence intervals are provided via bootstrapping.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://murphymv.github.io/semEff/}
  \item \url{https://github.com/murphymv/semEff}
  \item Report bugs at \url{https://github.com/murphymv/semEff/issues}
}

}
\author{
\strong{Maintainer}: Mark Murphy \email{murphymv@gmail.com}

}
\keyword{internal}
