% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Shipley.SEM.Boot}
\alias{Shipley.SEM.Boot}
\title{Bootstrapped SEM Estimates}
\format{A list of objects of class \code{"boot"}, representing bootstrapped
  estimates from fitted mixed models.}
\usage{
Shipley.SEM.Boot
}
\description{
Bootstrapped estimates generated from the hypothesised SEM from
  Shipley (2009), using \code{bootEff}.
}
\examples{
## Specification

\dontrun{

Shipley.SEM.Boot <- bootEff(Shipley.SEM, ran.eff = "site", seed = 53908)
}
}
\references{
Shipley, B. (2009). Confirmatory path analysis in a generalized
  multilevel context. \emph{Ecology}, \strong{90}(2), 363-368.
  \url{https://doi.org/bqd43d}
}
\keyword{datasets}
