% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shipley.sem.eff}
\alias{shipley.sem.eff}
\title{Effects for Shipley SEM}
\format{
A list object of class \code{"semEff"}, containing SEM effects and summary
tables.
}
\usage{
shipley.sem.eff
}
\description{
SEM effects calculated from bootstrapped estimates of the
hypothesised SEM from Shipley (2009), using \code{\link[=semEff]{semEff()}}.
}
\examples{
# Specification
# shipley.sem.eff <- semEff(shipley.sem.boot)
}
\references{
Shipley, B. (2009). Confirmatory path analysis in a generalized
multilevel context. \emph{Ecology}, \emph{90}(2), 363-368. \doi{10/bqd43d}
}
\keyword{datasets}
