% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Shipley.SEM.Eff}
\alias{Shipley.SEM.Eff}
\title{Effects for Shipley SEM}
\format{
A list object of class \code{"semEff"}, containing SEM effects and
  summary tables.
}
\usage{
Shipley.SEM.Eff
}
\description{
SEM effects calculated from bootstrapped estimates of the
  hypothesised SEM from Shipley (2009), using \code{semEff}.
}
\examples{
## Specification

\donttest{

Shipley.SEM.Eff <- semEff(Shipley.SEM.Boot)
}
}
\references{
Shipley, B. (2009). Confirmatory path analysis in a generalized
  multilevel context. \emph{Ecology}, \strong{90}(2), 363-368.
  \url{https://doi.org/bqd43d}
}
\keyword{datasets}
