% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdEff-fun.R
\name{stdCoeff}
\alias{stdCoeff}
\title{Standardised Effects}
\usage{
stdCoeff(...)
}
\arguments{
\item{...}{Arguments to \code{stdEff}.}
}
\description{
This is a temporary function which calls the renamed original
  (\code{\link[semEff]{stdEff}}).
}
